Update src/Jellyfin.Plugin.Dlna/Profiles/Xml/LG Smart TV.xml
Just tested this and it resolves both issues, with it set like this it picks up my lg tv as well as not picking up a device with an L in it. It does pick up if the device is named with lg in it like playlger, so theres probably some more work that needs done to make this better but this resolves picking up anything with just an L in the name.

Co-authored-by: Bond-009 <bond.009@outlook.com>